fix: separate component from constants to make hot reloading more effective
